Updated the code

+++ cpp-template.cpp
@@ -12,11 +12,11 @@ int main() {
   freopen("input.txt", "r", stdin);
   freopen("output.txt", "w", stdout);
 
-  lli testCase;
+  int testCase;
   cin >> testCase;
 
   while (testCase--) {
-    
+  	
   }
 
   return 0;
